+++ dotnet/samples/dev-team/DevTeam.AgentHost/Program.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Program.cs
+using Microsoft.AutoGen.Runtime.Grpc;
 
-using Microsoft.AutoGen.Agents;
 var builder = WebApplication.CreateBuilder(args);
 
 builder.AddServiceDefaults();
