+++ dotnet/src/Microsoft.AutoGen/Agents/HostBuilderExtensions.cs
@@ -116,47 +116,6 @@ public static class HostBuilderExtensions
         return property?.GetValue(null) as MessageDescriptor;
     }
 }
-public sealed class ReflectionHelper
-{
-    public static bool IsSubclassOfGeneric(Type type, Type genericBaseType)
-    {
-        while (type != null && type != typeof(object))
-        {
-            if (genericBaseType == (type.IsGenericType ? type.GetGenericTypeDefinition() : type))
-            {
-                return true;
-            }
-            if (type.BaseType == null)
-            {
-                return false;
-            }
-            type = type.BaseType;
-        }
-        return false;
-    }
-}
-public sealed class AgentTypes(Dictionary<string, Type> types)
-{
-    public Dictionary<string, Type> Types { get; } = types;
-    public static AgentTypes? GetAgentTypesFromAssembly()
-    {
-        var agents = AppDomain.CurrentDomain.GetAssemblies()
-                                .SelectMany(assembly => assembly.GetTypes())
-                                .Where(type => ReflectionHelper.IsSubclassOfGeneric(type, typeof(Agent))
-                                    && !type.IsAbstract
-                                    && !type.Name.Equals(nameof(Client)))
-                                .ToDictionary(type => type.Name, type => type);
-
-        return new AgentTypes(agents);
-    }
-}
-public sealed class EventTypes(TypeRegistry typeRegistry, Dictionary<string, Type> types, Dictionary<Type, HashSet<string>> eventsMap)
-{
-    public TypeRegistry TypeRegistry { get; } = typeRegistry;
-    public Dictionary<string, Type> Types { get; } = types;
-    public Dictionary<Type, HashSet<string>> EventsMap { get; } = eventsMap;
-}
-
 public sealed class AgentApplicationBuilder(IHostApplicationBuilder builder)
 {
     public AgentApplicationBuilder AddAgent<
