+++ dotnet/src/Microsoft.AutoGen/Runtime.Grpc/AgentWorkerHostingExtensions.cs
@@ -11,7 +11,7 @@ namespace Microsoft.AutoGen.Runtime.Grpc;
 
 public static class AgentWorkerHostingExtensions
 {
-    public static IHostApplicationBuilder AddAgentService(this IHostApplicationBuilder builder, bool local = false, bool useGrpc = true)
+    public static WebApplicationBuilder AddAgentService(this WebApplicationBuilder builder, bool local = false, bool useGrpc = true)
     {
         builder.AddOrleans(local);
 
@@ -25,11 +25,6 @@ public static class AgentWorkerHostingExtensions
         }
 
         return builder;
-    }
-
-    public static IHostApplicationBuilder AddLocalAgentService(this IHostApplicationBuilder builder, bool useGrpc = true)
-    {
-        return builder.AddAgentService(local: true, useGrpc);
     }
 
     public static WebApplication MapAgentService(this WebApplication app, bool local = false, bool useGrpc = true)
