+++ dotnet/src/Microsoft.AutoGen/Runtime.Grpc/Services/IGateway.cs
